Часть 12: Протокол OAuth (аутентификация через соц. сети)

+++ app/controllers/links_controller.rb
@@ -1,11 +1,12 @@
+# frozen_string_literal: true
+
+# LinksController
 class LinksController < ApplicationController
   before_action :authenticate_user!, only: :destroy
   before_action :link
 
   def destroy
-    if current_user&.author_of?(link.linkable)
-      link.destroy
-    end
+    link.destroy if current_user&.author_of?(link.linkable)
   end
 
   private
